Unify tests, fix lint errors
